Create F5_APM_RDP_SessionPersistence.irule
This iRule helps to load balance sessions for RDP servers in a pool. Basically you do not need to use any kind of iRule for pool have one rdp server. But in case of more than one server, you need to take care of session persistency, monitoring and load balancing.