Create APM_Policy_extractor.sh
This tools can be used to extract any APM policy stored in qkview or ucs file. However, the extracted context can only be imported with tmsh (as known as "load sys config from-terminal merge"). The extracting process creates two files which contains the APM policy and customizations. 

But I have to warn you, if you really need to export an APM policy, you better use the ucs file to import to some empty appliances and then export it from ACCESS policy web page. This is better because all F5 devices supports this way. 

If you use this script, you have to deal with all the missing dependencies and customizations which is very annoying. A couple of times i've used this script to export policies and i admit that is very hard.